gocode auto close

+++ liteidex/src/plugins/litebuild/litebuild.cpp
@@ -941,8 +941,9 @@ void LiteBuild::extFinish(bool error,int exitCode, QString msg)
 void LiteBuild::stopAction()
 {
     if (m_process->isRuning()) {
-        m_process->waitForFinished(100);
-        m_process->kill();
+        if (!m_process->waitForFinished(100)) {
+            m_process->kill();
+        }
     }
 }
